moved register and login components to their own folders
